// src/app/api/model/resultsetColumnHeaderData.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements. See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership. The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ResultsetColumnHeaderData { 
    booleanDisplayType?: boolean;
    codeLookupDisplayType?: boolean;
    codeValueDisplayType?: boolean;
    columnCode?: string;
    columnDisplayType?: ResultsetColumnHeaderData.ColumnDisplayTypeEnum;
    columnLength?: number;
    columnName?: string;
    columnType?: ResultsetColumnHeaderData.ColumnTypeEnum;
    columnValues?: Array<object>;
    dateDisplayType?: boolean;
    dateTimeDisplayType?: boolean;
    decimalDisplayType?: boolean;
    integerDisplayType?: boolean;
    isColumnIndexed?: boolean;
    isColumnNullable?: boolean;
    isColumnPrimaryKey?: boolean;
    isColumnUnique?: boolean;
    mandatory?: boolean;
    stringDisplayType?: boolean;
    textDisplayType?: boolean;
    timeDisplayType?: boolean;
}
export namespace ResultsetColumnHeaderData {
    export const ColumnDisplayTypeEnum = {
        Text: 'TEXT',
        String: 'STRING',
        Integer: 'INTEGER',
        Float: 'FLOAT',
        Decimal: 'DECIMAL',
        Date: 'DATE',
        Time: 'TIME',
        Datetime: 'DATETIME',
        Boolean: 'BOOLEAN',
        Binary: 'BINARY',
        Codelookup: 'CODELOOKUP',
        Codevalue: 'CODEVALUE'
    } as const;
    export type ColumnDisplayTypeEnum = typeof ColumnDisplayTypeEnum[keyof typeof ColumnDisplayTypeEnum];
    export const ColumnTypeEnum = {
        Bit: 'BIT',
        Boolean: 'BOOLEAN',
        Smallint: 'SMALLINT',
        Tinyint: 'TINYINT',
        Integer: 'INTEGER',
        Mediumint: 'MEDIUMINT',
        Bigint: 'BIGINT',
        Real: 'REAL',
        Float: 'FLOAT',
        Double: 'DOUBLE',
        Numeric: 'NUMERIC',
        Decimal: 'DECIMAL',
        Serial: 'SERIAL',
        Smallserial: 'SMALLSERIAL',
        Bigserial: 'BIGSERIAL',
        Money: 'MONEY',
        Char: 'CHAR',
        Varchar: 'VARCHAR',
        Longvarchar: 'LONGVARCHAR',
        Text: 'TEXT',
        Tinytext: 'TINYTEXT',
        Mediumtext: 'MEDIUMTEXT',
        Longtext: 'LONGTEXT',
        Json: 'JSON',
        Date: 'DATE',
        Time: 'TIME',
        TimeWithTimezone: 'TIME_WITH_TIMEZONE',
        Timestamp: 'TIMESTAMP',
        Datetime: 'DATETIME',
        TimestampWithTimezone: 'TIMESTAMP_WITH_TIMEZONE',
        Interval: 'INTERVAL',
        Binary: 'BINARY',
        Varbinary: 'VARBINARY',
        Longvarbinary: 'LONGVARBINARY',
        Bytea: 'BYTEA',
        Blob: 'BLOB',
        Tinyblob: 'TINYBLOB',
        Mediumblob: 'MEDIUMBLOB',
        Longblob: 'LONGBLOB'
    } as const;
    export type ColumnTypeEnum = typeof ColumnTypeEnum[keyof typeof ColumnTypeEnum];
}